Move filters to pkg/es
